Confirmation on Delete of Item in PO list Fixed.

+++ resources/js/purchase_order.js
@@ -185,20 +185,21 @@ $(document).ready(function() {
 
 function editPOItemRow(action,po_item_id) {
     if( 'delete' === action ){
-        alert('Are You Sure ?');
-        $.ajax({
-            type: "DELETE",
-            url: '/order/purchase/item/delete/'+po_item_id,
-            success: function(data)
-            {
-                if(true === data){
-                    alert('There was an error in Deleting this ROW. Please try again');
+        if(confirm('Are You Sure to Delete this Item?')){
+            $.ajax({
+                type: "DELETE",
+                url: '/order/purchase/item/delete/'+po_item_id,
+                success: function(data)
+                {
+                    if(true === data){
+                        alert('There was an error in Deleting this ROW. Please try again');
+                    }
+                    else{
+                        $("#"+po_item_id).remove();
+                    }
                 }
-                else{
-                    $("#"+po_item_id).remove();
-                }
-            }
-        });
+            });
+        }
     }
     else if( 'save' === action ){
